move some macro defines

+++ include/proc.hh
@@ -176,7 +176,7 @@ namespace proc
     };
     struct SleepingTask {
         struct Task *m_task;
-        uint m_wakeup_tick;
+        long m_wakeup_tick;
         SleepingTask():m_task(nullptr), m_wakeup_tick(0) {}
         SleepingTask(struct Task *a_task, uint a_tick):m_task(a_task), m_wakeup_tick(a_tick) {}
     };
